airflow/requirements.txt
apache-airflow==2.7.0
psycopg2-binary==2.9.9
clickhouse-driver==0.2.5
pandas==2.0.3